Delete project endpoint

+++ siem/service/src/main/java/bsep/sw/services/ProjectService.java
@@ -3,6 +3,7 @@ package bsep.sw.services;
 import bsep.sw.domain.Project;
 import bsep.sw.domain.User;
 import bsep.sw.repositories.ProjectRepository;
+import com.sun.org.apache.xpath.internal.operations.Bool;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -34,8 +35,13 @@ public class ProjectService {
         return repository.findOne(id);
     }
 
-    public void delete(final Long id) {
-        repository.delete(id);
+    public Boolean delete(final User user, final Long id) {
+        if (repository.findProjectByOwnerAndId(user, id) == null) {
+            return false;
+        } else {
+            repository.delete(id);
+            return true;
+        }
     }
 
     public Project findByUserAndId(final User user, final Long id) {
